Added in memory memcached stub for tests
